Move mapview to view folder
